add drafts for 2 new modules instead of one
