+++ bin/download_fasta_utils.py
@@ -1,15 +1,14 @@
 import csv
 import logging
-from ftplib import FTP
 import os
 import shutil
 import urllib.parse
-
+from ftplib import FTP
 
 import boto3
+import requests
 from botocore import UNSIGNED
 from botocore.config import Config
-import requests
 from retry import retry
 
 
@@ -18,8 +17,8 @@ def get_fasta_url(accession, analysis_ftp_field="generated_ftp"):
         file_url = f"https://api.ncbi.nlm.nih.gov/datasets/v2alpha/genome/accession/{accession}/download?include_annotation_type=GENOME_FASTA"
         return file_url
 
-    api_endpoint = 'https://www.ebi.ac.uk/ena/portal/api/search'
-    accession_type = 'analysis' if accession.startswith("ERZ") else 'wgs_set'
+    api_endpoint = "https://www.ebi.ac.uk/ena/portal/api/search"
+    accession_type = "analysis" if accession.startswith("ERZ") else "wgs_set"
     query = {
         "wgs_set": {
             "result": "wgs_set",
@@ -42,12 +41,12 @@ def get_fasta_url(accession, analysis_ftp_field="generated_ftp"):
         field_name = query[accession_type]["fields"]
         file_url = row[field_name].split(";")[0]  # Split to take the first FTP link if multiple
         return file_url
-    return None # no information about this accession in ENA
+    return None  # no information about this accession in ENA
 
 
 # TODO list errors explicitly, raise ValueError instead of returning None
-@retry(tries=5, delay=15, backoff=1.5) 
-def download_from_ENA_FIRE(accession: str, analysis_ftp_field: str, outpath: str) -> str:
+@retry(tries=5, delay=15, backoff=1.5)
+def download_from_ENA_FIRE(accession: str, analysis_ftp_field: str, outpath: str):
     url = get_fasta_url(accession, analysis_ftp_field=analysis_ftp_field)
     if not url:
         logging.debug(f"{accession} URL is empty for accession, ftp field: {analysis_ftp_field}")
@@ -70,18 +69,15 @@ def download_from_ENA_FIRE(accession: str, analysis_ftp_field: str, outpath: str
     # raise ValueError(f"Downloaded file {outpath} has zero size")
 
 
-@retry(tries=7, delay=15, backoff=2) 
+@retry(tries=7, delay=15, backoff=2)
 def download_from_ENA_API(accession: str, outpath: str) -> str:
     api_endpoint = f"https://www.ebi.ac.uk/ena/browser/api/fasta/{accession}"
     logging.debug(f"Download {accession} from ENA API using URL {api_endpoint}")
-    query = {
-        'download': 'true',
-        'gzip': 'true'
-    }
+    query = {"download": "true", "gzip": "true"}
     response = requests.get(api_endpoint, params=urllib.parse.urlencode(query))
     response.raise_for_status()
-    
-    with open(outpath, 'wb') as out:
+
+    with open(outpath, "wb") as out:
         out.write(response.content)
     # 20 bytes is a size of an empty fa.gz
     if os.path.exists(outpath) and os.path.getsize(outpath) > 20:
@@ -93,7 +89,7 @@ def download_from_ENA_API(accession: str, outpath: str) -> str:
 
 
 # TODO list errors explicitly, raise ValueError instead of returning None
-@retry(tries=8, delay=10, backoff=3) 
+@retry(tries=8, delay=10, backoff=3)
 def download_from_ENA_FTP(accession, outpath):
     url = get_fasta_url(accession)
     if not url:
@@ -101,13 +97,13 @@ def download_from_ENA_FTP(accession, outpath):
         return None
         # raise ValueError(f"URL is empty")
     logging.debug(f"Download {accession} from ENA FTP using URL {url}")
-    
+
     ftp_server = "ftp.ebi.ac.uk"
     ftp_path = url.replace(ftp_server, "")
-    
+
     with FTP(ftp_server) as ftp:
         ftp.login()
-        with open(outpath, 'wb') as file:
+        with open(outpath, "wb") as file:
             ftp.retrbinary(f"RETR {ftp_path}", file.write)
     # 20 bytes is a size of an empty fa.gz
     if os.path.exists(outpath) and os.path.getsize(outpath) > 20:
@@ -120,11 +116,11 @@ def download_from_ENA_FTP(accession, outpath):
 
 
 def download_from_NCBI_datasets(accession, download_folder):
-    outpath = os.path.join(download_folder, f'{accession}.fa')
+    outpath = os.path.join(download_folder, f"{accession}.fa")
     accession_version = accession if "." in accession else accession + ".1"
     api_endpoint = f"https://api.ncbi.nlm.nih.gov/datasets/v2alpha/genome/accession/{accession_version}/download"
     query = {
-        'include_annotation_type': 'GENOME_FASTA',
+        "include_annotation_type": "GENOME_FASTA",
     }
     response = run_request(query, api_endpoint)
 
@@ -132,12 +128,12 @@ def download_from_NCBI_datasets(accession, download_folder):
     tmp_archive = "ncbi_tmp.zip"
     tmp_archive_path = os.path.join(download_folder, tmp_archive)
     tmp_path = tmp_archive_path.replace(".zip", "")
-    with open(tmp_archive_path, 'wb') as out:
+    with open(tmp_archive_path, "wb") as out:
         out.write(content)
     shutil.unpack_archive(tmp_archive_path, tmp_path)
     subdir_path = os.path.join(download_folder, f"ncbi_tmp/ncbi_dataset/data/{accession_version}/")
     source_file = [file for file in os.listdir(subdir_path) if file.endswith("_genomic.fna")]
-    source_path = os.path.join(subdir_path, source_file[0]) # assembly_file is a list with one element
+    source_path = os.path.join(subdir_path, source_file[0])  # assembly_file is a list with one element
     shutil.move(source_path, outpath)
     os.remove(tmp_archive_path)
     shutil.rmtree(tmp_path)
