--- modules/local/find_assembly/tests/data/input_accessions.tsv
@@ -1,5 +1,0 @@
-ERZ1022847
-ERZ1462793
-GCA_938030915
-RSAC01
-CAMPAA01
