--- modules/local/find_assembly/tests/main.nf.test
@@ -1,27 +1,0 @@
-nextflow_process {
-
-    name "Test Process FIND_PRIMARY_ASSEMBLY"
-    script "../main.nf"
-    process "FIND_PRIMARY_ASSEMBLY"
-
-    test("Find assemblies for the test list of accessions") {
-
-        when {
-            params {
-                portion_size = 2
-            }
-            process {
-                """
-                input[0] = Channel.fromPath("${moduleDir}/tests/data/input_accessions.tsv", checkIfExists: true).splitText(by: params.portion_size, file: "portion")
-                """
-            }
-        }
-
-        then {
-            assert process.success
-            assert snapshot(process.out).match()
-        }
-
-    }
-
-}
